blob/pyproject.toml
[project]
name = "blob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "azure-identity>=1.25.0",
    "azure-storage-blob>=12.26.0",
    "fastapi[standard]>=0.116.1",
    "mimetype>=0.1.5",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
]
